Mongo Db Edit

+++ 02-MongoDb/02-Todos/src/controllers/todos.controllers.js
@@ -53,11 +53,50 @@ export const deleteTodos = async (req, res) => {
             return res.status(400).json({ message: "Invalid ID." });
         }
 
+        const isTodoFound = await todosModels.findById(id);
+
+        if (!isTodoFound) {
+            return res.status(404).json({ message: "Todo Not found." });
+        }
+
         await todosModels.findByIdAndDelete(id);
 
         res.status(201).json({
             message: "Deleted Successfully",
         })
+    } catch (error) {
+        console.log(error.message);
+        res.status(500).json({
+            message: "Internal Server Error",
+        });
+    };
+};
+
+export const editTodos = async (req, res) => {
+    try {
+        const { id } = req.params;
+        const { newTitle } = req.body;
+        if (!id) return res.status(400).json({
+            message: "id not found",
+        });
+
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({ message: "Invalid ID." });
+        }
+
+        const isTodoFound = await todosModels.findById(id);
+
+        if (!isTodoFound) {
+            return res.status(404).json({ message: "Todo Not found." });
+        }
+
+        const editedTodo = await todosModels.findByIdAndUpdate(id, { title: newTitle }, { new: true });
+
+        res.status(201).json({
+            message: "Edited Successfully",
+            editedTodo,
+        });
+
     } catch (error) {
         console.log(error.message);
         res.status(500).json({
